Add Spring Boot application deployment, service, and secret configurations; update Locust test setup and application endpoints

+++ 20_Application/springboot-app/src/main/kotlin/com/raynor/geek/springbootapp/service/CacheService.kt
@@ -19,7 +19,7 @@ class CacheService(
             redisTemplate.opsForValue().set(key, rand.toString(), 60, TimeUnit.SECONDS)
             rand.toString()
         }
-        return CacheResponseDto(key, value.toInt())
+        return CacheResponseDto(key, value)
     }
 }
